pyproject.toml
[tool.poetry]
name = "fastapi_boilerplate"
version = "0.1.0"
description = "Backend template"
authors = ["Mikhail Kraevskii <kramihm@gmail.com>"]
readme = "README.md"
packages = [{ include = "core" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100.0"
email-validator = "^2.0.0.post2"
orjson = "^3.9.1"
uvicorn = "^0.22.0"
python-dotenv = "^1.0.0"
asyncpg = "^0.28.0"
passlib = "^1.7.4"
python-jose = "^3.3.0"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pylint = "^2.17.4"
httpx = "^0.24.1"
pre-commit = "^3.3.3"
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
